+++ src/components/Calendar.tsx
@@ -1,11 +1,40 @@
 import React from "react";
 import styles from "./calendar.module.css";
-import { times, days, daysToStr } from "components/constants";
 import { useQuery } from "react-query";
 import { z } from "zod";
 import axios from "axios";
 import { usePageState } from "./state";
 import { StringDateSchema } from "../pages/subject";
+
+export const times = [
+  "",
+  "8:00",
+  "9:00",
+  "10:00",
+  "11:00",
+  "12:00",
+  "13:00",
+  "14:00",
+  "15:00",
+  "16:00",
+  "17:00",
+  "18:00",
+  "19:00",
+  "20:00",
+  "21:00",
+];
+
+export const days = ["su", "mo", "tu", "we", "th", "fr", "sa"] as const;
+
+export const daysToStr = {
+  su: "Sunday",
+  mo: "Monday",
+  tu: "Tuesday",
+  we: "Wednesday",
+  th: "Thursday",
+  fr: "Friday",
+  sa: "Saturday",
+} as const;
 
 const CalendarMeetingSchema = z.object({
   minutesDuration: z.number(),
